requirements.txt
annotated-types==0.7.0
anyio==4.5.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pytesseract==0.3.13
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
python-multipart
